requirements.txt
numpy==1.19.5
opencv-python-headless==4.5.5.62
Pillow==8.4.0
streamlit==1.5.0
tensorflow==2.7.0
tensorflow-hub==0.12.0
icrawler
faster-whisper
pyaudio
transformers
torch